Move research reward initialization code to Tally
This moves the procedure that initializes research reward context to the
tally. It improves start-up performance a bit and prepares the tally for
changes needed to support the switch to delta accrual tracking for block
version 11. This code does not belong in the transaction DB.
